Refactor resolve_dependencies() into session.c
Also keep track of session allocations and expose close_session(). Both
of these changes will be helpful in porting knit-resume-session to C.

+++ session.c
@@ -3,7 +3,22 @@
 #include "alloc.h"
 #include "hash.h"
 #include "job.h"
-#include "resource.h"
+
+static struct bytebuf session_bb;
+static struct bump_list* session_bump;
+
+static void* bmalloc(size_t size) {
+    size_t* raw = bump_alloc(&session_bump, sizeof(size) + size);
+    *raw = size;
+    return raw + 1;
+}
+
+static void* brealloc(void* ptr, size_t size) {
+    void* copy = bmalloc(size);
+    if (ptr)
+        memcpy(copy, ptr, ((size_t*)ptr)[-1]); // ptr is leaked
+    return copy;
+}
 
 #define MAX_SESSION_NUM (~(uint32_t)0)
 static_assert(MAX_SESSION_NUM <= SIZE_MAX);
@@ -11,6 +26,8 @@ static_assert(MAX_SESSION_NUM <= SIZE_MAX);
 struct session_step** active_steps;
 size_t num_active_steps;
 static size_t alloc_active_steps;
+
+size_t num_active_fanout;
 
 struct session_input** active_inputs;
 size_t num_active_inputs;
@@ -27,21 +44,21 @@ static void ensure_alloc_steps() {
     if (num_active_steps < alloc_active_steps)
         return;
     alloc_active_steps = realloc_size(alloc_active_steps);
-    active_steps = xrealloc(active_steps, alloc_active_steps * sizeof(*active_steps));
+    active_steps = brealloc(active_steps, alloc_active_steps * sizeof(*active_steps));
 }
 
 static void ensure_alloc_inputs() {
     if (num_active_inputs < alloc_active_inputs)
         return;
     alloc_active_inputs = realloc_size(alloc_active_inputs);
-    active_inputs = xrealloc(active_inputs, alloc_active_inputs * sizeof(*active_inputs));
+    active_inputs = brealloc(active_inputs, alloc_active_inputs * sizeof(*active_inputs));
 }
 
 static void ensure_alloc_deps() {
     if (num_active_deps < alloc_active_deps)
         return;
     alloc_active_deps = realloc_size(alloc_active_deps);
-    active_deps = xrealloc(active_deps, alloc_active_deps * sizeof(*active_deps));
+    active_deps = brealloc(active_deps, alloc_active_deps * sizeof(*active_deps));
 }
 
 size_t create_session_step(const char* name) {
@@ -51,7 +68,7 @@ size_t create_session_step(const char* name) {
         die("step created after fanout started");
 
     size_t namelen = strlen(name);
-    struct session_step* ss = xmalloc(sizeof(struct session_step) + namelen + 1);
+    struct session_step* ss = bmalloc(sizeof(struct session_step) + namelen + 1);
     memset(ss, 0, sizeof(*ss));
     if (namelen > SS_NAMEMASK)
         die("step name too long %s", name);
@@ -79,7 +96,7 @@ size_t create_session_input(size_t step_pos, const char* name) {
     }
 
     size_t namelen = strlen(name);
-    struct session_input* si = xmalloc(sizeof(struct session_input) + namelen + 1);
+    struct session_input* si = bmalloc(sizeof(struct session_input) + namelen + 1);
     memset(si, 0, sizeof(*si));
     si->step_pos = htonl(step_pos);
     if (namelen > SI_PATHMASK)
@@ -111,7 +128,7 @@ size_t create_session_dependency(size_t input_pos,
     ss_inc_unresolved(active_steps[dependent_pos]);
 
     size_t outlen = strlen(output);
-    struct session_dependency* sd = xmalloc(sizeof(struct session_dependency) + outlen + 1);
+    struct session_dependency* sd = bmalloc(sizeof(struct session_dependency) + outlen + 1);
     memset(sd, 0, sizeof(*sd));
     sd->input_pos = htonl(input_pos);
     sd->step_pos = htonl(step_pos);
@@ -126,8 +143,8 @@ size_t create_session_dependency(size_t input_pos,
     return num_active_deps++;
 }
 
-size_t num_active_fanout;
-
+// add_session_fanout_step() must be called after all normal steps have been
+// created; its returned step position is always greater than num_active_steps.
 size_t add_session_fanout_step() {
     if (num_active_steps + num_active_fanout == MAX_SESSION_NUM)
         die("too many fanout steps");
@@ -138,8 +155,11 @@ static char session_filepath[PATH_MAX];
 static char session_lockfile[PATH_MAX];
 static const char* session_name;
 
+static int unlock_session_atexit;
+
 static void unlock_session() {
-    unlink(session_lockfile);
+    if (*session_lockfile)
+        unlink(session_lockfile);
 }
 
 static int set_session_name(const char* sessname) {
@@ -156,9 +176,6 @@ static int set_session_name(const char* sessname) {
         session_name = session_filepath + len - strlen(sessname);
     }
 
-    if (snprintf(session_lockfile, PATH_MAX,
-                 "%s.lock", session_filepath) >= PATH_MAX)
-        return error("lock path too long");
     return 0;
 }
 
@@ -171,6 +188,10 @@ static int set_session_name_and_lock(const char* sessname) {
 
     if (set_session_name(sessname) < 0)
         return -1;
+
+    if (snprintf(session_lockfile, PATH_MAX,
+                 "%s.lock", session_filepath) >= PATH_MAX)
+        return error("lock path too long");
 
     int fd;
     long backoff_ms = 1;
@@ -186,7 +207,11 @@ static int set_session_name_and_lock(const char* sessname) {
         usleep(sleep_us);
     }
 
-    atexit(unlock_session);
+    if (!unlock_session_atexit) {
+        atexit(unlock_session);
+        unlock_session_atexit = 1;
+    }
+
     return 0;
 }
 
@@ -208,23 +233,22 @@ int new_session(const char* sessname) {
 }
 
 int load_current_session() {
-    struct bytebuf bb; // leaked
-    if (mmap_or_slurp_file(session_filepath, &bb) < 0)
+    if (mmap_or_slurp_file(session_filepath, &session_bb) < 0)
         return -1;
 
-    if (bb.size < sizeof(struct session_header))
+    if (session_bb.size < sizeof(struct session_header))
         return error("truncated session header");
-    struct session_header* hdr = bb.data;
+    struct session_header* hdr = session_bb.data;
 
     alloc_active_steps = num_active_steps = ntohl(hdr->num_steps);
-    active_steps = xmalloc(alloc_active_steps * sizeof(struct session_step*));
+    active_steps = bmalloc(alloc_active_steps * sizeof(struct session_step*));
     alloc_active_inputs = num_active_inputs = ntohl(hdr->num_inputs);
-    active_inputs = xmalloc(alloc_active_inputs * sizeof(struct session_input*));
+    active_inputs = bmalloc(alloc_active_inputs * sizeof(struct session_input*));
     alloc_active_deps = num_active_deps = ntohl(hdr->num_deps);
-    active_deps = xmalloc(alloc_active_deps * sizeof(struct session_dependency*));
+    active_deps = bmalloc(alloc_active_deps * sizeof(struct session_dependency*));
     num_active_fanout = ntohl(hdr->num_fanout);
 
-    char* end = (char*)bb.data + bb.size;
+    char* end = (char*)session_bb.data + session_bb.size;
     char* p = (char*)hdr + sizeof(*hdr);
     for (size_t i = 0; i < num_active_steps; i++) {
         struct session_step* ss = (struct session_step*)p;
@@ -272,6 +296,27 @@ int load_session_nolock(const char* sessname) {
     return load_current_session();
 }
 
+void close_session() {
+    active_steps = NULL;
+    num_active_steps = num_active_fanout = alloc_active_steps = 0;
+
+    active_inputs = NULL;
+    num_active_inputs = alloc_active_inputs = 0;
+
+    active_deps = NULL;
+    num_active_deps = alloc_active_deps = 0;
+
+    deps_dirty = 0;
+
+    free_bump_list(&session_bump);
+    cleanup_bytebuf(&session_bb);
+
+    unlock_session();
+
+    session_filepath[0] = session_lockfile[0] = '\0';
+    session_name = NULL;
+}
+
 static int cmp_dep(const void* a, const void* b) {
     const struct session_dependency* pa = *(const struct session_dependency**)a;
     const struct session_dependency* pb = *(const struct session_dependency**)b;
@@ -296,6 +341,7 @@ static int cmp_dep(const void* a, const void* b) {
 
 int save_session() {
     assert(session_name);
+    assert(*session_lockfile);
 
     if (deps_dirty)
         qsort(active_deps, num_active_deps, sizeof(*active_deps), cmp_dep);
@@ -435,4 +481,123 @@ int compile_job_for_step(size_t step_pos) {
     memcpy(ss->job_hash, job->object.oid.hash, KNIT_HASH_RAWSZ);
     ss_setflag(ss, SS_JOB);
     return 0;
+}
+
+static void set_input_resource(struct session_input* si, struct resource* res) {
+    memcpy(si->res_hash, res->object.oid.hash, KNIT_HASH_RAWSZ);
+    si_setflag(si, SI_RESOURCE | SI_FINAL);
+}
+
+// Create a session_input for each output with a matching prefix; outputs should
+// be positioned at the first match. The prefix will be removed from the input.
+static void create_fanout_inputs(size_t fanout_step_pos,
+                                 const struct resource_list* outputs,
+                                 const char* prefix) {
+    do {
+        // When depending on an entire production, skip special .knit/ outputs.
+        if (!*prefix && !strncmp(outputs->name, ".knit/", 6)) {
+            outputs = outputs->next;
+            continue;
+        }
+
+        char* name = outputs->name + strlen(prefix);
+        size_t input_pos = create_session_input(fanout_step_pos, name);
+        struct session_input* si = active_inputs[input_pos];
+        set_input_resource(si, outputs->res);
+        outputs = outputs->next;
+    } while (outputs && !strncmp(outputs->name, prefix, strlen(prefix)));
+}
+
+void resolve_dependencies(size_t step_pos, const struct resource_list* outputs) {
+    size_t dep_pos = 0;
+    while (dep_pos < num_active_deps && ntohl(active_deps[dep_pos]->step_pos) < step_pos)
+        dep_pos++;
+
+    while (dep_pos < num_active_deps) {
+        struct session_dependency* dep = active_deps[dep_pos];
+        if (dep->step_pos != htonl(step_pos))
+            break;
+
+        // Iterate production outputs in parallel with the dependencies waiting
+        // on this step. Compare their paths to find matching dependencies.
+        int cmp = 1;
+        if (outputs) {
+            cmp = sd_hasflag(dep, SD_PREFIX)
+                ? strncmp(outputs->name, dep->output, strlen(dep->output))
+                : strcmp(outputs->name, dep->output);
+        }
+
+        // If a production output has no dependencies on it, we can skip it.
+        if (cmp < 0) {
+            outputs = outputs->next;
+            continue;
+        }
+
+        struct session_input* input;
+        size_t dependent_pos;
+        if (sd_hasflag(dep, SD_INPUTISSTEP)) {
+            input = NULL;
+            dependent_pos = ntohl(dep->input_pos);
+            if (sd_hasflag(dep, SD_PREFIX))
+                die("SD_INPUTISSTEP and SD_PREFIX are incompatible");
+        } else {
+            size_t input_pos = ntohl(dep->input_pos);
+            if (input_pos >= num_active_inputs)
+                die("input out of bounds");
+            input = active_inputs[input_pos];
+            dependent_pos = ntohl(input->step_pos);
+        }
+        if (dependent_pos >= num_active_steps)
+            die("dependent out of bounds");
+        if (dependent_pos <= step_pos)
+            die("step later than its dependent");
+        struct session_step* dependent = active_steps[dependent_pos];
+
+        // Otherwise, we have a dependency to resolve. If we have no matching
+        // production output, then the dependency is missing.
+        if (cmp > 0) {
+            if (input)
+                si_setflag(input, SI_FINAL);
+            if (!sd_hasflag(dep, SD_REQUIRED)) {
+                goto mark_resolved;
+            } else if (!ss_hasflag(dependent, SS_FINAL)) {
+                // If the missing dependency is required, its dependent step
+                // must finish (with unmet requirements). Since there are no
+                // production outputs, we trivially resolve dependencies that
+                // are provided by the dependent step. This in turn may finish
+                // several additional steps with unmet requirements.
+                resolve_dependencies(dependent_pos, NULL);
+                ss_setflag(dependent, SS_FINAL);
+            }
+            dep_pos++;
+            continue;
+        }
+
+        // Note: cmp == 0
+        // The production and dependency outputs match.
+
+        // If the dependency is a prefix, create a fanout step whose inputs are
+        // all matching production outputs.
+        if (sd_hasflag(dep, SD_PREFIX)) {
+            assert(input);
+            size_t fanout_step_pos = add_session_fanout_step();
+            create_fanout_inputs(fanout_step_pos, outputs, dep->output);
+            si_setflag(input, SI_FANOUT | SI_FINAL);
+            input->fanout_step_pos = htonl(fanout_step_pos);
+        } else if (input) {
+            // Set a single matching resource for the input.
+            set_input_resource(input, outputs->res);
+        }
+
+mark_resolved:
+        if (!dependent->num_unresolved)
+            die("num_unresolved underflow on step %s", dependent->name);
+        ss_dec_unresolved(dependent);
+        if (!dependent->num_unresolved)
+            if (compile_job_for_step(dependent_pos) < 0)
+                exit(1);
+        dep_pos++;
+        // The same production output may resolve additional dependencies, so
+        // leave it for the next iteration.
+    }
 }
